% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_face_functions.R
\name{add_face_point}
\alias{add_face_point}
\title{Add a 69th point to face template}
\usage{
add_face_point(template, write_out = c("tem", "console"))
}
\arguments{
\item{write_out}{}
}
\value{
Face coordinates with 69 (instead of 68) face points. Can
be returned as a .tem for JPsychomorph.
}
\description{
Add a 69th point to face template
}
\details{
The 69th point is estimated to be at the top of the head,
somewhere between the hairline and back of the head. This is useful
if one is using dlib's 68 face point auto detector but also wants
an extra point to expand image size to include more of the face.

69th point is estimated by taking the face width (dlib points 2 & 14) and
multiplying it by 1.5.
}
