% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundness_angularity.R
\name{calc_angulrity}
\alias{calc_angulrity}
\title{Extract angularity of the face}
\usage{
calc_angulrity(image, landmarks = NULL)
}
\arguments{
\item{image}{Image to calculate race roundness}

\item{landmarks}{Optional. Landmarks provided via `read_landmarks` function.

#' @return Dataframe with angularity values for the provided image. See details
for what each named value corresponds with.}
}
\description{
Extract angularity of the face
}
