% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundness_angularity.R
\name{calc_roundness}
\alias{calc_roundness}
\title{Extract roundness of the face}
\usage{
calc_roundness(
  image,
  landmarks = NULL,
  point_select = "face_outline",
  fit = "taubin"
)
}
\arguments{
\item{image}{Image to calculate race roundness}

\item{landmarks}{Optional. Landmarks provided via `read_landmarks` function.}

\item{point_select}{One of `c("face_interior","face_outline", "face_hair","mouth","right_eye","left_eye")` or
vector of point landmarks (0-68) to fit the ellipse to. Defaults to `face_outline` which is most appropriate
for full face roundness measures in most cases.}

\item{fit}{One of `c("taubin","direct")` for algorithm to use for fitting ellipse to points. Defaults to `taubin`.}
}
\value{
Dataframe with roundness and circularity values for the provided image. See details
for what each named value corresponds with.
}
\description{
Extract roundness of the face
}
\details{
The `*1` returns indicate that the scale is set to 1; thus, the closer to 1 the value
is the more it is like a circle.
\itemize{
 \item{"roundness1}{Calculated as: `parimeter^2/(4*pi*area)`. Perfect circle = 1.}
 \item{"roundness2"}{Calculated as: `4*area/pi*R_major^2`. TBA}
 \item("circularity1"){Calculated as: `4*pi*(area/parimeter^2)`. Perfect circle = 1.}
 \item("circularity2"){Calculated as: `(4*pi*area)/parimeter`. TBA}
}
}
