% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fwhr.R
\name{calculate_fwhr}
\alias{calculate_fwhr}
\title{Calculate Facial Width-to-Height Ratio}
\usage{
calculate_fwhr(points, method = "average", top = "eyebrow")
}
\arguments{
\item{points}{Points returned by `read_landmarks` function}

\item{method}{either `average`, `left`, or `right` to determine which points on which side of the face to use for determining metrics.}

\item{top}{Determine top of face by either `eyebrow` or `eyelid`}
}
\value{

}
\description{
Calculate Facial Width-to-Height Ratio
}
