% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_face_functions.R
\name{create_average_template}
\alias{create_average_template}
\title{Create an average template from multiple templates}
\usage{
create_average_template(files)
}
\arguments{
\item{files}{List of files to be averaged.}
}
\value{

}
\description{
Create an average template from multiple templates
}
\examples{

}
