% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{extract_angularity}
\alias{extract_angularity}
\title{Extract face angularity}
\usage{
extract_angularity(
  image = image,
  landmarks = NULL,
  data_save_folder = data_save_folder,
  sqlite_db = sqlite_db,
  return_results = return_results
)
}
\arguments{
\item{return_results}{}
}
\value{

}
\description{
Extract face angularity
}
