% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{extract_hist_colors}
\alias{extract_hist_colors}
\title{Extract color histogram metric from interior of face}
\usage{
extract_hist_colors(
  image,
  data_save_folder = data_save_folder,
  sqlite_db = NULL,
  return_results = FALSE,
  bins = 32L
)
}
\arguments{
\item{return_results}{}
}
\value{

}
\description{
Extract color histogram metric from interior of face
}
