% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{extract_kmean_colors}
\alias{extract_kmean_colors}
\title{Extract kmeans color metric from interior face}
\usage{
extract_kmean_colors(
  image,
  data_save_folder,
  return_results = FALSE,
  kclusts = 25
)
}
\arguments{
\item{kclusts}{}
}
\value{

}
\description{
Extract kmeans color metric from interior face
}
