% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{extract_luminance}
\alias{extract_luminance}
\title{Extract luminance, RGB, and HSV values from cheeks and foreead of face}
\usage{
extract_luminance(image, data_save_folder, return_results = FALSE, buffer = 2)
}
\arguments{
\item{buffer}{}
}
\value{

}
\description{
Extract luminance, RGB, and HSV values from cheeks and foreead of face
}
