% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{extract_roundness}
\alias{extract_roundness}
\title{Extract face roundness}
\usage{
extract_roundness(
  image = image,
  landmarks = NULL,
  point_select = "face_outline",
  fit = "taubin",
  data_save_folder = data_save_folder,
  sqlite_db = sqlite_db,
  return_results = return_results
)
}
\arguments{
\item{return_results}{}
}
\value{

}
\description{
Extract face roundness
}
