% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{extract_structure}
\alias{extract_structure}
\title{Extract structure metric from interior of the face}
\usage{
extract_structure(
  image,
  data_save_folder = data_save_folder,
  sqlite_db = NULL,
  return_results = FALSE
)
}
\arguments{
\item{return_results}{}
}
\value{

}
\description{
Extract structure metric from interior of the face
}
