% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{extract_texture}
\alias{extract_texture}
\title{Extract texture metric from interior face}
\usage{
extract_texture(image, data_save_folder, return_results = FALSE)
}
\arguments{
\item{return_results}{}
}
\value{

}
\description{
Extract texture metric from interior face
}
