% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_face_functions.R
\name{face_crop_points_old}
\alias{face_crop_points_old}
\alias{Daniel}
\alias{N.}
\alias{Albohn}
\title{Crop a face and rescale}
\usage{
face_crop_points_old(
  landmarks,
  image,
  points = "default",
  savename,
  eye_dist = 125,
  rescale = TRUE,
  wh = 512,
  ...
)
}
\arguments{
\item{dist}{}
}
\value{

}
\description{
Crop a face and rescale
}
\examples{

}
