% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_face_functions.R
\name{face_hist}
\alias{face_hist}
\title{Extract color histogram from face image}
\usage{
face_hist(img, shape = c(8, 8, 8))
}
\arguments{
\item{shape}{}
}
\description{
Extract color histogram from face image
}
\examples{

}
