% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_face_functions.R
\name{face_mask}
\alias{face_mask}
\title{Mask an image to interior face only}
\usage{
face_mask(image, bb = NULL, crop = TRUE, savename,
  return_img = FALSE)
}
\arguments{
\item{return_img}{}
}
\description{
Mask an image to interior face only
}
