% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_face_functions.R
\name{face_swap}
\alias{face_swap}
\title{Swap faces between two images}
\usage{
face_swap(reference, target, convert = FALSE)
}
\arguments{
\item{reference}{Path to image with face to host swapped face from \code{target} image.}

\item{target}{Path to image with face to swap to \code{reference} image.}

\item{convert}{Logical. Convert returned image to be \code{R} compatible?
Defaults to \code{FALSE} and returns numpy matrix.}
}
\value{
Either a numpy matrix or EBImage object.
}
\description{
Swap faces between two images
}
