% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_face_functions.R
\name{landmarks_to_json}
\alias{landmarks_to_json}
\title{Converts landmarks file to json file}
\usage{
landmarks_to_json(landmarks, write_out = TRUE, savename)
}
\arguments{
\item{landmarks}{Landmarks returned from \link[faceplyr]{read_landmarks}}

\item{write_out}{If \code{TRUE} (default) write file to disk, else returns landmarks as json.}

\item{savename}{If supplied the name of the file to write out, else will derive directory from
landmarks file supplied.}
}
\value{
File or landmarks as type \code{.json}
}
\description{
Converts landmarks file to json file
}
\examples{
img <- system.file("extdata", "obama.png", package = "faceplyr")
landmarks <- read_landmarks(img)
landmarks_to_json(landmarks)

}
