% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_face_functions.R
\name{landmarks_to_tem}
\alias{landmarks_to_tem}
\title{Converts landmarks to JPsychomorph tem file}
\usage{
landmarks_to_tem(landmarks, write_out = TRUE, savename)
}
\arguments{
\item{landmarks}{Landmarks returned from \link[quantIm]{read_landmarks}}

\item{write_out}{If \code{TRUE} (default) write file to disk, else returns landmarks as json.}

\item{savename}{If supplied the name of the file to write out, else will derive directory from
landmarks file supplied.}
}
\value{
\code{.tem} file or data frame
}
\description{
Converts landmarks to JPsychomorph tem file
}
\examples{

}
