% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_face_functions.R
\name{tem_to_landmarks}
\alias{tem_to_landmarks}
\title{Convert JPsychomorph template to landmarks}
\usage{
tem_to_landmarks(tem, write_out = TRUE, savename)
}
\arguments{
\item{tem}{Template file}

\item{write_out}{Should the file be written out or returned?}

\item{savename}{File save name without extension (optional).}
}
\value{

}
\description{
Convert a \code{.tem} file created by JPsychomoprh back to a standard facial landmarks file.
}
\examples{

}
