% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_face_functions.R
\name{warp_face}
\alias{warp_face}
\title{Warp image with affine transformation}
\usage{
warp_face(image, m)
}
\arguments{
\item{image}{Path to/image to be warped.}

\item{m}{Affine matrix.}
}
\value{

}
\description{
Warp image with affine transformation
}
\examples{

}
