% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{extract_elements}
\alias{extract_elements}
\title{Wrapper function for metric extraction}
\usage{
extract_elements(
  image,
  data_save_folder = NULL,
  dbname = "face_features",
  return_results = FALSE,
  landmarks = NULL
)
}
\arguments{
\item{return_results}{}
}
\value{

}
\description{
Wrapper function for metric extraction
}
