% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{extract_elements}
\alias{extract_elements}
\title{Helper wrapper function for metric extraction}
\usage{
extract_elements(image, data_save_folder = "./data")
}
\arguments{
\item{data_save_folder}{}
}
\value{

}
\description{
Helper wrapper function for metric extraction
}
