% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_face_functions.R
\name{face_mask}
\alias{face_mask}
\title{Mask an image to interior face only}
\usage{
face_mask(
  image,
  crop = TRUE,
  savename = NULL,
  return_img = FALSE,
  transparent = TRUE
)
}
\arguments{
\item{transparent}{}
}
\value{

}
\description{
Mask an image to interior face only
}
