% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_face_functions.R
\name{face_texture}
\alias{face_texture}
\title{Extract Haralick texture features from face image}
\usage{
face_texture(img)
}
\arguments{
\item{img}{}
}
\value{

}
\description{
Extract Haralick texture features from face image
}
\examples{

}
