% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_landmarks.R
\name{read_landmarks}
\alias{read_landmarks}
\title{Read face landmarks}
\usage{
read_landmarks(x, ...)
}
\arguments{
\item{x}{Image or file to read. Acceptable files include JPsychomorph .tem files,
PNG, or JPG images.}

\item{...}{Additional arguments to pass (unused).}
}
\value{
Data frame of type \code{landmarks} with 68 facial landmarks.
}
\description{
Functions to read/obtain facial landmarks from files or images.
}
\examples{
img <- system.file("extdata", "obama.png", package = "faceplyr")
read_landmarks(img)
read_landmarks("template.tem")

}
